+++ azure/functions/decorators/function_app.py
@@ -43,7 +43,8 @@ from azure.functions.decorators.timer import TimerTrigger
 from azure.functions.decorators.utils import parse_singular_param_to_enum, \
     parse_iterable_param_to_enums, StringifyEnumJsonEncoder
 from azure.functions.http import HttpRequest
-from .generic import GenericInputBinding, GenericTrigger, GenericOutputBinding, ConnectorTrigger
+from .connectors import ConnectorTrigger
+from .generic import GenericInputBinding, GenericTrigger, GenericOutputBinding
 from .openai import _AssistantSkillTrigger, OpenAIModels, _TextCompletionInput, \
     _AssistantCreateOutput, \
     _AssistantQueryInput, _AssistantPostInput, InputType, _EmbeddingsInput, \
@@ -1572,8 +1573,6 @@ class TriggerApi(DecoratorApi, ABC):
 
         All optional fields will be given default values by the function host when
         they are parsed.
-
-        Ref: https://aka.ms/azure-function-binding-custom
 
         :param arg_name: The name of the trigger parameter in the function code.
         :param data_type: Defines how the Functions runtime should treat the
